webservice/requirements.txt
fastai==2.7.15
fastapi-slim[standard]==0.111.0